// pokemon/src/app/services/pokemon.interface.ts
export interface kanto {
  descriptions?: DescriptionsEntity[] | null;
  id: number;
  is_main_series: boolean;
  name: string;
  names?: NamesEntity[] | null;
  pokemon_entries?: PokemonEntriesEntity[] | null;
  region: LanguageOrPokemonSpeciesOrVersionGroupsEntityOrRegion;
  version_groups?:
    | LanguageOrPokemonSpeciesOrVersionGroupsEntityOrRegion[]
    | null;
}
export interface DescriptionsEntity {
  description: string;
  language: LanguageOrPokemonSpeciesOrVersionGroupsEntityOrRegion;
}
export interface LanguageOrPokemonSpeciesOrVersionGroupsEntityOrRegion {
  name: string;
  url: string;
}
export interface NamesEntity {
  language: LanguageOrPokemonSpeciesOrVersionGroupsEntityOrRegion;
  name: string;
}
export interface PokemonEntriesEntity {
  entry_number: number;
  pokemon_species: LanguageOrPokemonSpeciesOrVersionGroupsEntityOrRegion;
}

export interface Pokemon {
  abilities: Ability[];
  base_experience: number;
  cries: Cries;
  forms: Species[];
  game_indices: GameIndex[];
  height: number;
  held_items: any[];
  id: number;
  is_default: boolean;
  location_area_encounters: string;
  moves: Move[];
  name: string;
  order: number;
  past_abilities: PastAbility[];
  past_types: any[];
  species: Species;
  sprites: Sprites;
  stats: Stat[];
  types: Type[];
  weight: number;
}

export interface Ability {
  ability: Species | null;
  is_hidden: boolean;
  slot: number;
}

export interface Species {
  name: string;
  url: string;
}

export interface Cries {
  latest: string;
  legacy: string;
}

export interface GameIndex {
  game_index: number;
  version: Species;
}

export interface Move {
  move: Species;
  version_group_details: VersionGroupDetail[];
}

export interface VersionGroupDetail {
  level_learned_at: number;
  move_learn_method: Species;
  order: number | null;
  version_group: Species;
}

export interface PastAbility {
  abilities: Ability[];
  generation: Species;
}

export interface GenerationV {
  'black-white': Sprites;
}

export interface GenerationIv {
  'diamond-pearl': Sprites;
  'heartgold-soulsilver': Sprites;
  platinum: Sprites;
}

export interface Versions {
  'generation-i': GenerationI;
  'generation-ii': GenerationIi;
  'generation-iii': GenerationIii;
  'generation-iv': GenerationIv;
  'generation-v': GenerationV;
  'generation-vi': { [key: string]: Home };
  'generation-vii': GenerationVii;
  'generation-viii': GenerationViii;
}

export interface Other {
  dream_world: DreamWorld;
  home: Home;
  'official-artwork': OfficialArtwork;
  showdown: Sprites;
}

export interface Sprites {
  back_default: string;
  back_female: null;
  back_shiny: string;
  back_shiny_female: null;
  front_default: string;
  front_female: null;
  front_shiny: string;
  front_shiny_female: null;
  other?: Other;
  versions?: Versions;
  animated?: Sprites;
}
// ...
export interface Stat {
  base_stat: number;
  effort: number;
  stat: Species;
}

export interface Type {
  slot: number;
  type: Species;
}

export interface expectPokemon {
  picture: string;
  name: string;
  types: string[];
}